requirements.cpu.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.9.1+cpu
torchvision==0.10.1+cpu
torchaudio==0.9.0
tensorboard==2.5.0
tqdm==4.61.2
imageio==2.9.0
scikit-image==0.18.2
pyaml==20.4.0
imutils==0.5.4
ipykernel==6.0.1
opencv-python==4.5.3.56
joblib==1.0.1
scikit-learn==0.24.2
ptflops==0.6.6
pytz==2021.1
albumentations==1.0.3
pandas==1.3.1
lmdb==1.2.1
gdown==4.0.2